### 2021-09-13 #3717 @maxkfranz
The [suggested DOIs](https://github.com/openjournals/joss-reviews/issues/3717#issuecomment-917360026) look correct to me.  Thanks
