### 2021-09-14 #3717 @galessiorob
@Kevin-Mattheus-Moerman Happy to take this one!

@maxkfranz I'll start reaching out to [potential reviewers](https://bit.ly/joss-reviewers), but if you have any specific suggestions feel free to ping them here.
