### 2021-09-14 #3717 @maxkfranz
> @maxkfranz if the suggested DOIs look good you can start adding them to the actual doc.

I thought this was automatic.  Assuming you mean the bib file, I'll amend that.  Thanks

> @maxkfranz I'll start reaching out to potential reviewers, but if you have any specific suggestions feel free to ping them here.

Here are some suggestions:

- John Bachman @johnbachman
- Ben Gyori https://scholar.harvard.edu/bgyori/home
- Alex Pico @AlexanderPico
- Andrew Su @andrewsu
- Scooter Morris @scootermorris
