### 2021-09-14 b49c005
Add DOIs to paper.bib entries
