+++ paper.bib
@@ -3,7 +3,8 @@
   author={Wong, Jeffrey V and Franz, Max and Siper, Metin Can and Fong, Dylan and Durupinar, Funda and Dallago, Christian and Luna, Augustin and Giorgi, John M and Rodchenkov, Igor and Babur, {\"O}zg{\"u}n and others},
   journal={bioRxiv},
   year={2021},
-  publisher={Cold Spring Harbor Laboratory}
+  publisher={Cold Spring Harbor Laboratory},
+  doi={10.1101/2021.03.10.382333}
 }
 
 @article{cerami2010pathway,
@@ -14,7 +15,8 @@
   number={suppl\_1},
   pages={D685--D690},
   year={2010},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/nar/gkq1039}
 }
 
 @article{gyori2017word,
@@ -24,7 +26,8 @@
   volume={13},
   number={11},
   pages={954},
-  year={2017}
+  year={2017},
+  doi={10.15252/msb.20177651}
 }
 
 @article{pafilis2009reflect,
@@ -33,7 +36,8 @@
   journal={Nature Precedings},
   pages={1--1},
   year={2009},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
+  doi={10.1038/npre.2009.3212.1}
 }
 
 @article{berriz2008synergizer,
@@ -44,7 +48,8 @@
   number={19},
   pages={2272--2273},
   year={2008},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/bioinformatics/btn424}
 }
 
 @article{mering2003string,
@@ -55,7 +60,8 @@
   number={1},
   pages={258--261},
   year={2003},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/nar/gkg034}
 }
 
 @article{mostafavi2008genemania,
@@ -66,7 +72,8 @@
   number={1},
   pages={1--15},
   year={2008},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1186/gb-2008-9-s1-s4}
 }
 
 @article{reimand2007g,
@@ -77,7 +84,8 @@
   number={suppl\_2},
   pages={W193--W200},
   year={2007},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/nar/gkm226}
 }
 
 @article{cote2007protein,
@@ -88,7 +96,8 @@
   number={1},
   pages={1--14},
   year={2007},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1186/1471-2105-8-401}
 }
 
 @article{juty2012identifiers,
@@ -99,7 +108,8 @@
   number={D1},
   pages={D580--D586},
   year={2012},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/nar/gkr1097}
 }
 
 @article{abeysooriya2021gene,
@@ -107,7 +117,8 @@
   author={Abeysooriya, Mandhri and Soria, Megan and Kasu, Mary Sravya and Ziemann, Mark},
   journal={bioRxiv},
   year={2021},
-  publisher={Cold Spring Harbor Laboratory}
+  publisher={Cold Spring Harbor Laboratory},
+  doi={10.1101/2021.03.30.437702}
 }
 
 @misc{groundingswagger,
@@ -127,5 +138,6 @@
   number={suppl\_1},
   pages={D52--D57},
   year={2010},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/nar/gki031}
 }
