### 2021-09-23 #3756 @whedon
[REVIEW]: A flexible search system for high-accuracy identification of biological entities and molecules
**Submitting author:** @maxkfranz (<a href="http://orcid.org/0000-0003-0169-0480">Max Franz</a>)
**Repository:** <a href="https://github.com/PathwayCommons/grounding-search" target ="_blank">https://github.com/PathwayCommons/grounding-search</a>
**Version:** 0.6.0
