### 2021-09-23 3261ea0
Add mention of grounding.baderlab.org in readme
A more convenient alternative for testing than spinning up a local instance
