### 2021-09-27 #3756 @galessiorob
@AlexanderPico and @rabdill A checklist has been provided for each of you at the top of the issue [here.](https://github.com/openjournals/joss-reviews/issues/3756#issue-1005000830) Please let me know if you have any questions or any of the items or the paper itself.

Feel free to ping the author @maxkfranz with specific questions on the paper and the package here, or if you find it easier, you can document them on individual issues and mention them there.

Thanks again for volunteering 🎉
