### 2021-10-08 #3756 @AlexanderPico
- Haven't been successful in getting it installed and running via docker or npm yet. Still working on it...
  - Current error: `http://localhost:9200/groundingsearch => connect ECONNREFUSED 127.0.0.1:9200`
- Disclosing COI: I'm a co-author with Max Franz on a consortium paper that has been submitted for review. We are two of many, many authors representing our distinct contributions to Cytoscape. I do not feel I have an ethical conflict in this case and would request a waiver.
- State of the field: the paper is lacking a comparison with other related tools. A contrast with common ID mapping tools would be useful as well.
