### 2021-10-12 #3756 @apcamargo
I finished my review. On top of the (lack of) ticks, my only minor comment is that the species names in the manuscript are not italicized.
