### 2021-10-12 438a8aa
Added docker memory tip
