### 2021-10-25 1ed6f8b
Update readme with usage examples and an initial tool comparison
