### 2021-11-02 #3756 @galessiorob
👋  @AlexanderPico @rabdill @apcamargo

Could you give the paper the last review and check if the latest additions and edits satisfy the last items from the list, please? I believe that the "state of the field" and "statement of need" items are taken care of in the statement of need and discussion sections.

Thank you so much for your time and effort on this project.
