### 2021-11-02 #3756 @AlexanderPico
My remaining unchecked box is for "state of the field" under the "Software paper" heading. While I agree with the author that a tool comparison is best placed in the repo's README for easy updating, I'm not sure it satisfies the journal's requirement for the paper. I'm fine with either:

 - The editor's approval that the repo's README counts as part of the paper, or
 - A sentence being added to the paper stating that a tool comparison is available in the repo's README.
