### 2021-11-02 #3756 @apcamargo
I agree that the "state of the field" was taken care of. I still think that the README would be much improved if it included a better real-world usage example of the tool. It is very abstract right now.

Also, the scientific names are still not italicized.
