### 2021-11-03 70c4fac
Updates to paper & readme
- Italicised org. names
- Reference readme re. tool comparison
- Example video of Biofactoid in readme

Ref. :

- https://github.com/openjournals/joss-reviews/issues/3756#issuecomment-957072473
- https://github.com/openjournals/joss-reviews/issues/3756#issuecomment-958009531
- https://github.com/openjournals/joss-reviews/issues/3756
