+++ paper.md
@@ -63,11 +63,11 @@ bibliography: paper.bib
 
 # Summary
 
-Identifying subcellular biological entities (genes, gene products, and small molecules) is essential in using and creating bioinformatics analysis tools, text mining, and accessible biological research apps. When research information is uniquely and unambiguously identified, it enables data to be accurately retrieved, cross-referenced, and integrated. In practice, biological entities are identified when they are associated with a matching record from a knowledge base that specialises in collecting and organising information of that type (e.g. genes in NCBI Gene). Our search service increases the efficiency and ease of use for identifying biological entities compared to prior approaches [@reimand2007g; @cote2007protein; @berriz2008synergizer; @juty2012identifiers]. This identification service may be used to power research apps and tools, such as Biofactoid, GeneMANIA, and STRING, where colloquial entity names may be provided as input by users [@wong2021capturing; @mostafavi2008genemania; @mering2003string].
+Identifying subcellular biological entities (genes, gene products, and small molecules) is essential in using and creating bioinformatics analysis tools, text mining, and accessible biological research apps. When research information is uniquely and unambiguously identified, it enables data to be accurately retrieved, cross-referenced, and integrated. In practice, biological entities are identified when they are associated with a matching record from a knowledge base that specialises in collecting and organising information of that type (e.g. genes in NCBI Gene). Our search service increases the efficiency and ease of use for identifying biological entities compared to prior approaches [@reimand2007g; @cote2007protein; @berriz2008synergizer; @juty2012identifiers]. A comparison of similar tools is available in the software documentation[@githubRepoReadme]. This identification service may be used to power research apps and tools, such as Biofactoid, GeneMANIA, and STRING, where colloquial entity names may be provided as input by users [@wong2021capturing; @mostafavi2008genemania; @mering2003string].
 
 # Statement of need
 
-Most biologists are unaware of the concept of grounding data to database identifiers (i.e., a value, usually a piece of text, that uniquely identifies the entity) [@abeysooriya2021gene]. When an author labels a biological entity as ‘IL6’, for instance, they may not consider that this label could be ambiguous. Is this IL6 for Homo sapiens or for Mus musculus? If for Mus musculus, there is more than one gene that is called by that name. By mapping the user’s entity to a database identifier, in this case, 3569 in NCBI Gene, the data becomes disambiguated.
+Most biologists are unaware of the concept of grounding data to database identifiers (i.e., a value, usually a piece of text, that uniquely identifies the entity) [@abeysooriya2021gene]. When an author labels a biological entity as ‘IL6’, for instance, they may not consider that this label could be ambiguous. Is this IL6 for *Homo sapiens* or for *Mus musculus*? If for *Mus musculus*, there is more than one gene that is called by that name. By mapping the user’s entity to a database identifier, in this case, 3569 in NCBI Gene, the data becomes disambiguated.
 
 Because many biologists are unaware of the utility of database identifiers, they often have the perception that the common or canonical name of an entity is sufficient for use with analysis tools. These users can be confused by traditional database grounding interfaces, where it is the user’s responsibility to select a particular identifier from a long list manually. This sort of grounding interface is incongruent with users’ mental model: What purpose could this list of entities have when the entity has already been identified by name?
 
@@ -81,7 +81,7 @@ To verify the accuracy of the service’s results, a test suite was created. The
 
 # Mechanism
 
[…]
