+++ paper.bib
@@ -130,6 +130,15 @@
   url = {https://grounding.baderlab.org}
 }
 
+@misc{githubRepoReadme,
+  author = {Max Franz and Jeffrey V Wong and Metin Can Siper},
+  title = {GitHub: Pathway Commons Grounding Search Service},
+  year = {2021},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/PathwayCommons/grounding-search}
+}
+
 @article{gene2pubmed,
   title={Entrez Gene: gene-centered information at NCBI},
   author={Maglott, Donna and Ostell, Jim and Pruitt, Kim D and Tatusova, Tatiana},
