### 2021-11-03 #3756 @maxkfranz
Changes in latest proof:

- Brief reference to readme re. tool comparison
- Italicised organism names
- [Video example](https://github.com/PathwayCommons/grounding-search#description) of Biofactoid use case in the readme
