### 2021-11-09 #3756 @galessiorob
@AlexanderPico @apcamargo and @rabdill Thank you so much for lending your expertise and reviewing this paper 🚀

@maxkfranz The next thing we need from you is to deposit your software in Zenodo and post the DOI here. After that, we can set the final version, generate a final proof, and have an Editor in Chief do the publishing. 🙏
