### 2021-11-12 #3756 @kyleniemeyer
Hi @maxkfranz, can you please edit the metadata of the Zenodo archive so that the author list matches that of the paper?
