### 2021-11-12 #3756 @maxkfranz
The metadata on Zenodo has been updated: https://zenodo.org/record/4974637
